+++ CdIts.Caffoa.Cli/Config/CaffoaGlobalConfig.cs
@@ -11,5 +11,6 @@ public class CaffoaGlobalConfig : CaffoaConfig
         CheckEnums = true;
         UseDateOnly = false;
         Imports = new List<string>();
+        UseInheritance = true; 
     }
 }