+++ CdIts.Caffoa.Cli/Errors/CaffoaParserError.cs
@@ -5,4 +5,7 @@ public class CaffoaParserError : Exception
     public CaffoaParserError(string? message) : base(message)
     {
     }
+    public CaffoaParserError(string? message, Exception inner) : base(message, inner)
+    {
+    }
 }